download_sysext: integrate signature verification
Move the delta update part from test/crau_verify.rs to
update-format-crau.
Move verify_sig into update-format-crau, so it can be imported by
delta_update.
Integrate signature verification in download_sysext.
