+++ Cargo.toml
@@ -11,7 +11,6 @@ tokio = { version = "1", features = ["full"] }
 uuid = "1.2"
 sha2 = "0.10"
 url = "2"
-rsa = { version = "0.9.2", features = ["sha2"] }
 
 env_logger = "0.10"
 log = "0.4"
