+++ src/lib.rs
@@ -2,4 +2,3 @@ mod download;
 pub use download::download_and_hash;
 
 pub mod request;
-pub mod verify_sig;
