+++ update-format-crau/Cargo.toml
@@ -8,3 +8,4 @@ edition = "2021"
 [dependencies]
 log = "0.4.19"
 protobuf = "3"
+rsa = { version = "0.9.2", features = ["sha2"] }
