+++ update-format-crau/src/lib.rs
@@ -1,4 +1,6 @@
+pub mod delta_update;
 mod generated;
+pub mod verify_sig;
 
 pub mod proto {
     pub use super::generated::update_metadata::*;
